+++ payments-core/src/test/java/com/stripe/android/networking/StripeApiRepositoryTest.kt
@@ -10,7 +10,6 @@ import com.stripe.android.FinancialConnectionsFixtures
 import com.stripe.android.FraudDetectionDataFixtures
 import com.stripe.android.FraudDetectionDataRepository
 import com.stripe.android.Stripe
-import com.stripe.android.StripeCashAppPayBetaApi
 import com.stripe.android.core.exception.APIConnectionException
 import com.stripe.android.core.exception.InvalidRequestException
 import com.stripe.android.core.model.StripeFileParams
@@ -655,7 +654,6 @@ internal class StripeApiRepositoryTest {
                 )
         }
 
-    @OptIn(StripeCashAppPayBetaApi::class)
     @Test
     fun confirmPaymentIntent_withPaymentMethodCreateParamsAttribution_setsCorrectPaymentUserAgent() =
         runTest {
@@ -766,7 +764,6 @@ internal class StripeApiRepositoryTest {
                 )
         }
 
-    @OptIn(StripeCashAppPayBetaApi::class)
     @Test
     fun confirmSetupIntent_withPaymentMethodCreateParamsAttribution_setsCorrectPaymentUserAgent() =
         runTest {