+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/TestAuthorization.kt
@@ -1,6 +1,5 @@
 package com.stripe.android
 
-import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.test.espresso.intent.Intents
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -12,37 +11,26 @@ import com.stripe.android.test.core.Browser
 import com.stripe.android.test.core.Currency
 import com.stripe.android.test.core.Customer
 import com.stripe.android.test.core.DelayedPMs
-import com.stripe.android.test.core.DisableAnimationsRule
 import com.stripe.android.test.core.GooglePayState
-import com.stripe.android.test.core.INDIVIDUAL_TEST_TIMEOUT_SECONDS
 import com.stripe.android.test.core.IntentType
 import com.stripe.android.test.core.LinkState
 import com.stripe.android.test.core.MyScreenCaptureProcessor
 import com.stripe.android.test.core.PlaygroundTestDriver
 import com.stripe.android.test.core.Shipping
 import com.stripe.android.test.core.TestParameters
-import com.stripe.android.test.core.TestWatcher
+import com.stripe.android.utils.TestRules
 import com.stripe.android.utils.initializedLpmRepository
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.Timeout
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
 class TestAuthorization {
-    @get:Rule
-    var globalTimeout: Timeout = Timeout.seconds(INDIVIDUAL_TEST_TIMEOUT_SECONDS)
 
     @get:Rule
-    val composeTestRule = createEmptyComposeRule()
-
-    @get:Rule
-    val testWatcher = TestWatcher()
-
-    @get:Rule
-    val disableAnimations = DisableAnimationsRule()
+    val rules = TestRules.create()
 
     private lateinit var device: UiDevice
     private lateinit var testDriver: PlaygroundTestDriver
@@ -51,7 +39,7 @@ class TestAuthorization {
     @Before
     fun before() {
         device = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
-        testDriver = PlaygroundTestDriver(device, composeTestRule, screenshotProcessor)
+        testDriver = PlaygroundTestDriver(device, rules.compose, screenshotProcessor)
     }
 
     @After