+++ paymentsheet-example/src/androidTestDebug/java/com/stripe/android/TestMultiStepFieldsReloaded.kt
@@ -1,7 +1,6 @@
 package com.stripe.android
 
 import androidx.compose.ui.test.assertContentDescriptionEquals
-import androidx.compose.ui.test.junit4.createEmptyComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -13,39 +12,27 @@ import com.stripe.android.test.core.Billing
 import com.stripe.android.test.core.Currency
 import com.stripe.android.test.core.Customer
 import com.stripe.android.test.core.DelayedPMs
-import com.stripe.android.test.core.DisableAnimationsRule
 import com.stripe.android.test.core.GooglePayState
-import com.stripe.android.test.core.INDIVIDUAL_TEST_TIMEOUT_SECONDS
 import com.stripe.android.test.core.IntentType
 import com.stripe.android.test.core.LinkState
 import com.stripe.android.test.core.MyScreenCaptureProcessor
 import com.stripe.android.test.core.PlaygroundTestDriver
 import com.stripe.android.test.core.Shipping
 import com.stripe.android.test.core.TestParameters
-import com.stripe.android.test.core.TestWatcher
 import com.stripe.android.ui.core.forms.resources.LpmRepository
+import com.stripe.android.utils.TestRules
 import com.stripe.android.utils.initializedLpmRepository
 import org.junit.Before
 import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
-import org.junit.rules.Timeout
 import org.junit.runner.RunWith
-
 
 @RunWith(AndroidJUnit4::class)
 class TestMultiStepFieldsReloaded {
-    @get:Rule
-    var globalTimeout: Timeout = Timeout.seconds(INDIVIDUAL_TEST_TIMEOUT_SECONDS)
 
     @get:Rule
-    val composeTestRule = createEmptyComposeRule()
-
-    @get:Rule
-    val testWatcher = TestWatcher()
-
-    @get:Rule
-    val disableAnimations = DisableAnimationsRule()
+    val rules = TestRules.create()
 
     private lateinit var device: UiDevice
     private lateinit var testDriver: PlaygroundTestDriver
@@ -71,7 +58,7 @@ class TestMultiStepFieldsReloaded {
     @Before
     fun before() {
         device = UiDevice.getInstance(InstrumentationRegistry.getInstrumentation())
-        testDriver = PlaygroundTestDriver(device, composeTestRule, screenshotProcessor)
+        testDriver = PlaygroundTestDriver(device, rules.compose, screenshotProcessor)
     }
 
     @Test
@@ -103,14 +90,14 @@ class TestMultiStepFieldsReloaded {
                 authorizationAction = null
             ),
             populateCustomLpmFields = {
-                composeTestRule.onNodeWithText("IBAN").apply {
+                rules.compose.onNodeWithText("IBAN").apply {
                     performTextInput(
                         "DE89370400440532013000"
                     )
                 }
             },
             verifyCustomLpmFields = {
-                composeTestRule.onNodeWithText("IBAN").apply {
+                rules.compose.onNodeWithText("IBAN").apply {
                     assertContentDescriptionEquals(
                         "DE89370400440532013000"
                     )