+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/flowcontroller/FlowControllerViewModel.kt
@@ -1,15 +1,24 @@
 package com.stripe.android.paymentsheet.flowcontroller
 
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
 import com.stripe.android.model.ElementsSessionParams
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.state.PaymentSheetState
 
 internal class FlowControllerViewModel(
-    private val handle: SavedStateHandle
-) : ViewModel() {
+    application: Application,
+    private val handle: SavedStateHandle,
+) : AndroidViewModel(application) {
+
+    val flowControllerStateComponent: FlowControllerStateComponent =
+        DaggerFlowControllerStateComponent
+            .builder()
+            .appContext(application)
+            .flowControllerViewModel(this)
+            .build()
 
     var initializationMode: PaymentSheet.InitializationMode? = null