+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/PaymentOptionsActivityTest.kt
@@ -22,7 +22,6 @@ import com.google.common.truth.Truth.assertThat
 import com.stripe.android.ApiKeyFixtures
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.Logger
-import com.stripe.android.core.injection.DUMMY_INJECTOR_KEY
 import com.stripe.android.core.injection.Injectable
 import com.stripe.android.core.injection.NonFallbackInjector
 import com.stripe.android.core.injection.WeakMapInjectorRegistry
@@ -60,7 +59,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -109,13 +107,6 @@ internal class PaymentOptionsActivityTest {
     }
 
     private val viewModel = createViewModel()
-
-    private val primaryButtonUIState = PrimaryButton.UIState(
-        label = "Test",
-        onClick = {},
-        enabled = true,
-        visible = true
-    )
 
     private val ActivityPaymentOptionsBinding.continueButton: PrimaryButton
         get() = root.findViewById(R.id.primary_button)
@@ -298,82 +289,6 @@ internal class PaymentOptionsActivityTest {
     }
 
     @Test
-    fun `ContinueButton should be enabled when primaryButtonEnabled is true`() {
-        val scenario = activityScenario()
-        scenario.launch(
-            createIntent()
-        ).use {
-            it.onActivity { activity ->
-                viewModel.updatePrimaryButtonUIState(
-                    primaryButtonUIState.copy(
-                        enabled = true
-                    )
-                )
-                assertThat(activity.viewBinding.continueButton.isEnabled).isTrue()
-            }
-        }
-    }
-
-    @Test
-    fun `ContinueButton should be disabled when primaryButtonEnabled is false`() {
-        val scenario = activityScenario()
-        scenario.launch(
-            createIntent()
-        ).use {
-            it.onActivity { activity ->
-                viewModel.updatePrimaryButtonUIState(
-                    primaryButtonUIState.copy(
-                        enabled = false
-                    )
-                )
-                assertThat(activity.viewBinding.continueButton.isEnabled).isFalse()
-            }
-        }
-    }
-
-    @Test
-    fun `ContinueButton text should update when primaryButtonText updates`() {
-        val scenario = activityScenario()
-        scenario.launch(
-            createIntent()
-        ).use {
-            it.onActivity { activity ->
-                viewModel.updatePrimaryButtonUIState(
-                    primaryButtonUIState.copy(
-                        label = "Some text"
-                    )
-                )
-                assertThat(activity.viewBinding.continueButton.externalLabel).isEqualTo("Some text")
-            }
-        }
-    }
-
-    @Test
-    fun `ContinueButton should go back to initial state after updating selection`() {
-        Dispatchers.setMain(testDispatcher)
-
-        val scenario = activityScenario()
-        scenario.launch(
-            createIntent()
-        ).use {
-            it.onActivity { activity ->
-                viewModel.updatePrimaryButtonUIState(
-                    primaryButtonUIState.copy(
-                        label = "Some text",
-                        enabled = false
-                    )
-                )
-                assertThat(activity.viewBinding.continueButton.externalLabel).isEqualTo("Some text")
-                assertThat(activity.viewBinding.continueButton.isEnabled).isFalse()
-
-                viewModel.updateSelection(mock<PaymentSelection.New.Card>())
-                assertThat(activity.viewBinding.continueButton.externalLabel).isEqualTo("Continue")
-                assertThat(activity.viewBinding.continueButton.isEnabled).isTrue()
-            }
-        }
-    }
-
-    @Test
     fun `notes visibility is set correctly`() {
         val scenario = activityScenario()
         scenario.launch(
@@ -500,7 +415,6 @@ internal class PaymentOptionsActivityTest {
                 workContext = testDispatcher,
                 application = ApplicationProvider.getApplicationContext(),
                 logger = Logger.noop(),
-                injectorKey = DUMMY_INJECTOR_KEY,
                 lpmResourceRepository = StaticLpmResourceRepository(lpmRepository),
                 addressResourceRepository = StaticAddressResourceRepository(addressRepository),
                 savedStateHandle = savedStateHandle,